backend/requirements.txt
certifi==2018.4.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.0.7
django-cors-headers==2.4.0
django-filter==2.0.0
django-js-asset==1.1.0
django-mptt==0.9.0
django-templated-mail==1.1.1
django-versatileimagefield==1.9
djangorestframework==3.8.2
djoser==1.2.0
idna==2.7
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
openapi-codec==1.3.2
Pillow==5.0.0
pytz==2018.5
requests==2.19.1
simplejson==3.16.0
uritemplate==3.0.0
urllib3==1.23
